[TASK] Remove require statements for classes
Since the autoloader is in place, the require statements for
all classes (including template.php) is not necessary.

This patch adds the core classes to the autoloader file,
and removes all unnecessary require statements.
This change also brings modification with that moves the
initialization of $GLOBALS['TBE_TEMPLATE'] from template.php
into init.php. This means that TBE_TEMPLATE is always
available in the backend now.

Change-Id: I0759d678d42d122e0dc99568e761a05f4b036af4
Resolves: #38131
Reviewed-on: http://review.typo3.org/12142
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter