[TASK] Unify references for backend modules icons
Somehow on earth, regular (non-extbase) modules
register their module icons via [labels][tabs_images][tab]
when using addModule(). The icon is not related to the
labels at all, and can be simplified.

All icons registered are available via the module
configuration option "icon" and old formats are migrated
automatically (with a deprecation warning), and always
stored in an absolute format.

Using "icon" now in the available places makes it easier
and is unified in ModuleLoader, FormEngine, EXT:about and the
Module Menu.

Resolves: #72827
Releases: master
Change-Id: Ie5770539bcb94921a9319a1d8d53d981c49fbcd8
Reviewed-on: https://review.typo3.org/46100
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>