[!!!][TASK] Drop PageTS "TSFE.constants" functionality
When parsing TypoScript, PageTSconfig and UserTSconfig
is taking into account for evaluting if "TSFE.constants" is
set in PageTSconfig. If this is used, this acts as a default
constant.

However, this feature has various drawbacks, and has been
superseded by actually having constants globally (FE/BE) and via
Conditions available.

For frontend calls with empty pagesection caches, this improves
performance greatly, especially in pages on a deeper level (recursion).

This also reduces the footprint to have the TypoScriptParser (incl.
condition matching) being called too often.

Resolves: #88564
Releases: master
Change-Id: Icbdfa9918a9a79510a87198bf033c4de96d71107
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61034
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>