[TASK] Keep pid for versioned records
DataHandler does not set "-1" for versioned records in workspaces
anymore, thus making it a lot easier to find the real PID for a versioned
record without always having to fetch the live version for it.

This is ground-breaking, as this opens the door for a lot of improvements
when accessing versions via overlays.

An upgrade wizard will migrate all pid=-1 records to their equivalent
pids, and discard all already published / archived versions.

Resolves: #89555
Releases: master
Change-Id: I31c32451827c1f94764bb0ba22ad8207c8b3d4fd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61699
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Manuel Selbach <manuel_selbach@yahoo.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Manuel Selbach <manuel_selbach@yahoo.de>
Reviewed-by: Benni Mack <benni@typo3.org>