[TASK] Highlight invalid backend layout value
In #89700, the page information view in the info
module was extended for a new type, which displays
the layout related (backend_layout, ..) fields for the
corresponding pages.

Since there might be field values which are not longer
valid, because the referenced backend layout does not
longer exist, the new type does  now highlight such values
with the already known `INVALID VALUE ("<value>")` string.

Resolves: #94674
Releases: master
Change-Id: I342fff7b8d76694831fa839901250b7d5da7e973
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70194
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Simon Gilli <typo3@gilbertsoft.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>