[!!!][TASK] Move pages.fe_login_mode into custom extension
The pages.fe_login_mode field was initially introduced
to allow reducing the amount of possible cache variants.

Since this functionality is rarely used nowadays it is
now removed from Core, reducing complexity in TSFE.

The functionality will be made available as an extension,
which can be installed using a new upgrade wizard.

As a consequence, also the TypoScript config
"sendCacheHeaders_onlyWhenLoginDeniedInBranch"
is removed.

Resolves: #96616
Releases: main
Change-Id: I13609e87bc7c2332a029e51edd70f0205d218002
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73106
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>