[BUGFIX] Prevent overwriting of view variable
The InfoModuleController previously added the form
action url as "moduleName" to the view. Since #96730,
this variable name is overwritten by the ModuleTemplate.

Since then, instead of the form action url, the variable
only contained the module identifier. This lead to
misbehavior on submitting the form.

The form action url is therefore is now added, using
the dedicated "formAction" variable.

Resolves: #96798
Related: #96730
Releases: main
Change-Id: Ic3c737cf1168bf2d6e887cc77d4d2a301037c9df
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73368
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>